refactor(tasks): split the tests into seperate files
